+++ src/main/java/com/synopsys/integration/detect/configuration/DetectConfigurationFactory.java
@@ -286,8 +286,7 @@ public class DetectConfigurationFactory {
         String prefix = detectConfiguration.getNullableValue(DetectProperties.DETECT_PROJECT_CODELOCATION_PREFIX);
         String suffix = detectConfiguration.getNullableValue(DetectProperties.DETECT_PROJECT_CODELOCATION_SUFFIX);
         Boolean useBdio2 = detectConfiguration.getValue(DetectProperties.DETECT_BDIO2_ENABLED);
-        Boolean useLegacyUpload = detectConfiguration.getValue(DetectProperties.BLACKDUCK_LEGACY_UPLOAD_ENABLED);
-        return new BdioOptions(useBdio2, prefix, suffix, useLegacyUpload);
+        return new BdioOptions(useBdio2, prefix, suffix);
     }
 
     public ProjectNameVersionOptions createProjectNameVersionOptions(String sourceDirectoryName) {