+++ src/main/java/com/synopsys/integration/detect/workflow/bdio/BdioOptions.java
@@ -4,13 +4,11 @@ public class BdioOptions {
     private final boolean enabledBdio2;
     private final String projectCodeLocationSuffix;
     private final String projectCodeLocationPrefix;
-    private final boolean enabledLegacyUpload;
 
-    public BdioOptions(boolean enabledBdio2, String projectCodeLocationPrefix, String projectCodeLocationSuffix, boolean enabledLegacyUpload) {
+    public BdioOptions(boolean enabledBdio2, String projectCodeLocationPrefix, String projectCodeLocationSuffix) {
         this.enabledBdio2 = enabledBdio2;
         this.projectCodeLocationSuffix = projectCodeLocationSuffix;
         this.projectCodeLocationPrefix = projectCodeLocationPrefix;
-        this.enabledLegacyUpload = enabledLegacyUpload;
     }
 
     public String getProjectCodeLocationSuffix() {
@@ -24,8 +22,5 @@ public class BdioOptions {
     public boolean isBdio2Enabled() {
         return enabledBdio2;
     }
-
-    public boolean isLegacyUploadEnabled() {
-        return enabledLegacyUpload;
-    }
+    
 }